[FRAGMENTS] - Merge adapters - fix ux
