+++ app/src/main/res/layout/fragment_main.xml
@@ -17,6 +17,8 @@
             android:id="@+id/cardHeaderMainView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_margin="5dp"
+            app:cardCornerRadius="15dp"
             app:cardElevation="4dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
